Bump to stable pager API
moar was renamed to moor:
https://github.com/walles/moor/pull/305

As part of that change, a new stable public API was added.

This change bumps the pager API to moor v2.

I dropped the ShouldFormat setting, because that happens only for JSON
input. There should be no visible change.

I also dropped picking the style. That's done automatically based on
the terminal background color. For dark terminal backgrounds, "native"
is the default, so there should be no visible change.